chore: move files to sharre folder
